crime-service/requirements.txt
fastapi
uvicorn
asyncpg
sqlalchemy
python-dotenv
pytz
email_validator
passlib[bcrypt]==1.7.4
shortuuid==1.0.13
python-jose[cryptography]
redis==5.2.1
httpx 
pandas
scikit-learn
numpy
googlemaps
folium
xlrd>=2.0.1